Moved helper functions from words.js to wordHelpers.js and fixed the bug that caused the passTurn event call to not clear the gameboard from selections.
