+++ server/server_modules/Codenames/GameLogic/gameSetup.js
@@ -1,3 +1,4 @@
+// @ts-check
 const {
     updateTeamOrder,
     updateUser,
